requirements.txt
httpx>=0.27.0
supabase>=2.0.0
python-dotenv>=1.0.0
google-genai>=1.0.0
